+++ src/main/java/io/xlorey/FluxLoader/annotations/Injected.java
@@ -1,7 +1,10 @@
 package io.xlorey.FluxLoader.annotations;
 
+import java.lang.annotation.*;
+
 /**
  * Annotation - a flag indicating that injection occurred in this file
  */
+@Target({ElementType.METHOD})
 public @interface Injected {
 }