+++ src/main/java/io/xlorey/FluxLoader/annotations/SubscribeSingleEvent.java
@@ -1,12 +1,12 @@
 package io.xlorey.FluxLoader.annotations;
 
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.*;
 
 /**
  * Subscription to user events
  */
 @Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.METHOD})
 public @interface SubscribeSingleEvent {
     /**
      * Event name